Updated document formatting
Specifically, I set the automatic paragraph indent to be 0, as the
indentation was causing sections of the document to behave in a
strange manner.

In addition, I redefined the \indent command to perform an indent
the same size as the default automatic indentation. We probably
will not need to use the \indent command in this document, but it's
there for utility.